Fix Vim status line errors when cursor is on empty line
This error is given by Vim 9.0 when BQN status line is active
and cursor is moved over an empty line:

    Error detected while processing function bqn#l:
    line    2:
    E33: No previous substitute regular expression

This seems to happen when the result of substitution is empty,
i.e. the pattern didn't match anything, or the expansion was
the empty string.

Fix by appending a space to the line so substitute() always
matches something.